requirements.txt
blinker==1.6.2
click==8.1.6
dnspython==2.4.1
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
pymongo==4.4.1
Werkzeug==2.3.6
